Let MMS create the server user, fix REST publish not returning item ID.
Before when an app is created in MMX, the server user will be created.
Now with MMS integration, the server user will be created by MMS.  MMX
just binds the server user with the app.

REST publish API returned the IQ id which was not useful.  It should
return the item ID.  This affects v1 and v2.