Fix the missing appID in the from JID and creator JID of topic.
When sending a message from console using REST API, the <message from="serveruser-JID"...>
was missing the "%appID", so the intercept failed to route the message.
Also when a topic (or channel) was created the console, the creator was missing the
"%appID" in the JID..
Also support private channel created from console.