Use recipientUserIds in send msg JSON, and add send msg by user names.
V1 message send request used "recipientUsernames" for userID's, now it is
corrected.
Also message send by user names is supported, but no message will be sent
if any recipients in the request are invalid.  It is because the initial
conversion from user names to user ID's already validates the names.